+++ zcash_client_memory/src/testing/mod.rs
@@ -66,11 +66,16 @@ impl DataStoreFactory for TestMemDbFactory {
 #[derive(Debug)]
 pub struct MemBlockCacheInsertionResult {
     txids: Vec<TxId>,
+    compact_block: CompactBlock,
 }
 
 impl CacheInsertionResult for MemBlockCacheInsertionResult {
     fn txids(&self) -> &[TxId] {
         &self.txids[..]
+    }
+
+    fn compact_block(&self) -> Option<&CompactBlock> {
+        Some(&self.compact_block)
     }
 }
 
@@ -86,7 +91,10 @@ impl TestCache for MemBlockCache {
     fn insert(&mut self, cb: &CompactBlock) -> Self::InsertResult {
         let txids = cb.vtx.iter().map(|tx| tx.txid()).collect();
         self.0.write().unwrap().insert(cb.height(), cb.clone());
-        MemBlockCacheInsertionResult { txids }
+        MemBlockCacheInsertionResult {
+            txids,
+            compact_block: cb.clone(),
+        }
     }
 
     fn truncate_to_height(&mut self, height: BlockHeight) {